Update dummy app models to newest README.md
