+++ aaf-gumboot.gemspec
@@ -24,6 +24,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rake', '~> 10.0'
   spec.add_development_dependency 'mysql2'
   spec.add_development_dependency 'sqlite3'
+  spec.add_development_dependency 'valhammer'
   spec.add_development_dependency 'factory_girl'
   spec.add_development_dependency 'faker'
   spec.add_development_dependency 'rspec-rails', '~> 3.1'
