+++ spec/dummy/app/models/api_subject_role.rb
@@ -1,0 +1,6 @@
+class APISubjectRole < ActiveRecord::Base
+  belongs_to :api_subject
+  belongs_to :role
+
+  validates :api_subject, :role, presence: true
+end
