+++ spec/dummy/app/models/role.rb
@@ -1,5 +1,5 @@
 class Role < ActiveRecord::Base
-  has_many :api_subject_roles, class_name: 'API::APISubjectRole'
+  has_many :api_subject_roles
   has_many :api_subjects, through: :api_subject_roles
 
   has_many :subject_roles
@@ -7,5 +7,5 @@ class Role < ActiveRecord::Base
 
   has_many :permissions
 
-  validates :name, presence: true
+  valhammer
 end
