+++ spec/gumboot/api_subjects_spec.rb
@@ -2,6 +2,6 @@ require 'spec_helper'
 
 require 'gumboot/shared_examples/api_subjects'
 
-RSpec.describe API::APISubject, type: :model do
+RSpec.describe APISubject, type: :model do
   include_examples 'API Subjects'
 end
